Use Windows Native Dns Resolver
