+++ resolver-dns-native-windows/src/main/java/io/netty/resolver/dns/windows/WindowsAdapterInfo.java
@@ -1,0 +1,45 @@
+package io.netty.resolver.dns.windows;
+
+import io.netty.util.internal.NativeLibraryLoader;
+import io.netty.util.internal.PlatformDependent;
+
+public class WindowsAdapterInfo {
+    private static final Throwable UNAVAILABILITY_CAUSE;
+
+    static {
+        Throwable cause = null;
+        try {
+            loadNativeLibrary();
+        } catch (Throwable error) {
+            cause = error;
+        }
+        UNAVAILABILITY_CAUSE = cause;
+    }
+
+    private static void loadNativeLibrary() {
+        if (!PlatformDependent.isWindows()) {
+            throw new IllegalStateException("Only supported on Windows");
+        }
+
+        String sharedLibName = "netty_resolver_dns_native_windows" + '_' + PlatformDependent.normalizedArch();
+        ClassLoader cl = PlatformDependent.getClassLoader(WindowsAdapterInfo.class);
+        NativeLibraryLoader.load(sharedLibName, cl);
+    }
+
+    public static boolean isAvailable() {
+        return UNAVAILABILITY_CAUSE == null;
+    }
+
+    public static void ensureAvailability() {
+        if (UNAVAILABILITY_CAUSE != null) {
+            throw (Error) new UnsatisfiedLinkError(
+                    "failed to load the required native library").initCause(UNAVAILABILITY_CAUSE);
+        }
+    }
+
+    public static Throwable unavailabilityCause() {
+        return UNAVAILABILITY_CAUSE;
+    }
+
+    public static native NetworkAdapter[] adapters();
+}
