+++ resolver-dns/src/main/java/io/netty/resolver/dns/DnsNameResolver.java
@@ -143,7 +143,7 @@ public class DnsNameResolver extends InetNameResolver {
         String[] searchDomains;
         try {
             List<String> list = PlatformDependent.isWindows()
-                    ? getSearchDomainsHack()
+                    ? WindowsResolverDnsServerAddressStreamProvider.getSearchDomains()
                     : UnixResolverDnsServerAddressStreamProvider.parseEtcResolverSearchDomains();
             searchDomains = list.toArray(new String[0]);
         } catch (Exception ignore) {
@@ -183,23 +183,6 @@ public class DnsNameResolver extends InetNameResolver {
             // ignore
         }
         return false;
-    }
-
-    @SuppressWarnings("unchecked")
-    private static List<String> getSearchDomainsHack() throws Exception {
-        // Only try if not using Java9 and later
-        // See https://github.com/netty/netty/issues/9500
-        if (PlatformDependent.javaVersion() < 9) {
-            // This code on Java 9+ yields a warning about illegal reflective access that will be denied in
-            // a future release. There doesn't seem to be a better way to get search domains for Windows yet.
-            Class<?> configClass = Class.forName("sun.net.dns.ResolverConfiguration");
-            Method open = configClass.getMethod("open");
-            Method nameservers = configClass.getMethod("searchlist");
-            Object instance = open.invoke(null);
-
-            return (List<String>) nameservers.invoke(instance);
-        }
-        return Collections.emptyList();
     }
 
     private static final DatagramDnsResponseDecoder DATAGRAM_DECODER = new DatagramDnsResponseDecoder() {
