+++ resolver-dns/src/main/java/io/netty/resolver/dns/DnsServerAddressStreamProviders.java
@@ -94,7 +94,6 @@ public final class DnsServerAddressStreamProviders {
     /**
      * A {@link DnsServerAddressStreamProvider} which inherits the DNS servers from your local host's configuration.
      * <p>
-     * Note that only macOS and Linux are currently supported.
      * @return A {@link DnsServerAddressStreamProvider} which inherits the DNS servers from your local host's
      * configuration.
      */
@@ -110,6 +109,7 @@ public final class DnsServerAddressStreamProviders {
                 // ignore
             }
         }
+
         return unixDefault();
     }
 
@@ -123,8 +123,6 @@ public final class DnsServerAddressStreamProviders {
         // We use 5 minutes which is the same as what OpenJDK is using in sun.net.dns.ResolverConfigurationImpl.
         private static final long REFRESH_INTERVAL = TimeUnit.MINUTES.toNanos(5);
 
-        // TODO(scott): how is this done on Windows? This may require a JNI call to GetNetworkParams
-        // https://msdn.microsoft.com/en-us/library/aa365968(VS.85).aspx.
         static final DnsServerAddressStreamProvider DEFAULT_DNS_SERVER_ADDRESS_STREAM_PROVIDER =
                 new DnsServerAddressStreamProvider() {
                     private volatile DnsServerAddressStreamProvider currentProvider = provider();
@@ -145,9 +143,7 @@ public final class DnsServerAddressStreamProviders {
                     }
 
                     private DnsServerAddressStreamProvider provider() {
-                        // If on windows just use the DefaultDnsServerAddressStreamProvider.INSTANCE as otherwise
-                        // we will log some error which may be confusing.
-                        return PlatformDependent.isWindows() ? DefaultDnsServerAddressStreamProvider.INSTANCE :
+                        return PlatformDependent.isWindows() ? WindowsResolverDnsServerAddressStreamProvider.loadConfig() :
                                 UnixResolverDnsServerAddressStreamProvider.parseSilently();
                     }
                 };
